fixed middleware directory
