+++ backend/supabase_client.py
@@ -2,7 +2,6 @@ import os
 from flask import g
 from werkzeug.local import LocalProxy
 from supabase import create_client, ClientOptions
-from flask_storage import FlaskSessionStorage
 
 SUPABASE_URL = os.getenv("SUPABASE_URL", "")
 SUPABASE_KEY = os.getenv("SUPABASE_KEY", "")
@@ -13,7 +12,6 @@ def get_supabase():
             SUPABASE_URL,
             SUPABASE_KEY,
             options=ClientOptions(
-                storage=FlaskSessionStorage(),
                 flow_type="pkce",
                 auto_refresh_token=True
             ),
